refactor: extract the headless surface to internal/presentation/headless (#1087)
cmd/headless.go drops from 417 lines to cobra wiring: flags in, headless.Run
out. runHeadless becomes headless.Run, headlessOptions becomes headless.Options,
and the stdin IPC control loop moves alongside it as control.go.

inheritedSubagentMode was shared with cmd/chat.go, so it becomes
scheddomain.InheritedAgentMode next to the EnvSubagentAgentMode it reads.
